call orca directly
