+++ server/py/framework/utils/clients/base_client.py
@@ -13,26 +13,91 @@
 # limitations under the License.
 
 import asyncio
+import contextlib
+import copy
 import typing
 from abc import ABC, abstractmethod
 
+import aiohttp
 import fastapi
 from fastapi.concurrency import run_in_threadpool
 
 import mlrun.common.schemas
 import mlrun.errors
 import mlrun.utils.singleton
+from mlrun.utils import logger
 
 
 class BaseClient(ABC, metaclass=mlrun.utils.singleton.AbstractSingleton):
     def __init__(self, *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
-        self._api_url = mlrun.mlconf.iguazio_api_url
 
     @property
     def is_sync(self) -> bool:
         """Indicates whether the client is synchronous."""
         return True
+
+    @property
+    def _api_url(self) -> str:
+        return mlrun.mlconf.iguazio_api_url
+
+    @property
+    @abstractmethod
+    def _verify_session_http_method(self) -> str:
+        pass
+
+    @abstractmethod
+    def _generate_auth_info_from_session_verification_response(
+        self,
+        response_headers: typing.Mapping[str, typing.Any],
+        response_body: typing.Mapping[typing.Any, typing.Any],
+    ) -> mlrun.common.schemas.AuthInfo:
+        pass
+
+    @abstractmethod
+    def _prepare_request_kwargs(
+        self, session: typing.Optional[str], path: str, *, kwargs: dict
+    ):
+        pass
+
+    def _send_request_to_api(
+        self,
+        method: str,
+        path: str,
+        error_message: str,
+        session=None,
+        retry_on_post=False,
+        **kwargs,
+    ):
+        url = f"{self._api_url}/api/{path}"
+        self._prepare_request_kwargs(session, path, kwargs=kwargs)
+        http_session = self._session
+        if retry_on_post and self._retry_on_post_session:
+            http_session = self._retry_on_post_session
+        response = http_session.request(
+            method, url, verify=mlrun.mlconf.httpdb.http.verify, **kwargs
+        )
+        if not response.ok:
+            try:
+                response_body = response.json()
+            except Exception:
+                response_body = {}
+            self._handle_error_response(
+                method, path, response, response_body, error_message, kwargs
+            )
+        return response
+
+    @abstractmethod
+    def _handle_error_response(
+        self,
+        method: str,
+        path: str,
+        response: typing.Any,  # or aiohttp.ClientResponse
+        response_body: dict,
+        error_message: str,
+        kwargs: dict,
+    ) -> None:
+        pass
 
 
 class BaseAsyncClient(BaseClient):
@@ -42,11 +107,15 @@ class BaseAsyncClient(BaseClient):
         self._async_session: typing.Optional[mlrun.utils.AsyncClientWithRetry] = None
 
     @property
-    def is_sync(self):
+    def is_sync(self) -> bool:
         """
         False because client is asynchronous
         """
         return False
+
+    @property
+    def _session_verification_endpoint(self) -> str:
+        return mlrun.mlconf.httpdb.authentication.iguazio.session_verification_endpoint
 
     def __getattribute__(self, name):
         """
@@ -72,8 +141,87 @@ class BaseAsyncClient(BaseClient):
 
         return wrapper
 
-    @abstractmethod
     async def verify_request_session(
         self, request: fastapi.Request
     ) -> mlrun.common.schemas.AuthInfo:
-        pass
+        """
+        Proxy the request to one of the session verification endpoints (which will verify the session of the request)
+        """
+        headers = {
+            "authorization": request.headers.get("authorization"),
+            "cookie": request.headers.get("cookie", ""),
+            "x-request-id": getattr(request.state, "request_id", ""),
+        }
+        async with (
+            self._send_request_to_api_async(
+                self._verify_session_http_method,
+                self._session_verification_endpoint,
+                "Failed verifying iguazio session",
+                retry_options_override=mlrun.utils.async_http.ExponentialRetryOverride(
+                    blacklisted_methods=[],  # iguazio session verification endpoint is idempotent
+                    # 1, 2, 4, 8, ...
+                    start_timeout=1,
+                    max_timeout=30.0,
+                    factor=2.0,
+                ),
+                headers=headers,
+            ) as response
+        ):
+            return self._generate_auth_info_from_session_verification_response(
+                response.headers, await response.json()
+            )
+
+    @contextlib.asynccontextmanager
+    async def _send_request_to_api_async(
+        self,
+        method: str,
+        path: str,
+        error_message: str,
+        session: typing.Optional[str] = None,
+        retry_options_override: typing.Optional[
+            mlrun.utils.async_http.ExponentialRetryOverride
+        ] = None,
+        **kwargs,
+    ) -> typing.AsyncGenerator[aiohttp.ClientResponse, None]:
+        url = f"{self._api_url}/api/{path}"
+        self._prepare_request_kwargs(session, path, kwargs=kwargs)
+        await self._ensure_async_session()
+
+        # take the session default
+        retry_options = copy.deepcopy(self._async_session.retry_options)
+
+        # override with cherry-picked options
+        if retry_options_override:
+            if retry_options_override.blacklisted_methods is not None:
+                retry_options.blacklisted_methods = (
+                    retry_options_override.blacklisted_methods
+                )
+            retry_options._start_timeout = retry_options_override._start_timeout
+            retry_options._max_timeout = retry_options_override._max_timeout
+            retry_options._factor = retry_options_override._factor
+
+        response = None
+        try:
+            response = await self._async_session.request(
+                method, url, verify_ssl=False, retry_options=retry_options, **kwargs
+            )
+            if not response.ok:
+                try:
+                    response_body = await response.json()
+                except Exception:
+                    response_body = {}
+                self._handle_error_response(
+                    method, path, response, response_body, error_message, kwargs
+                )
+            yield response
+        finally:
+            if response:
+                response.release()
+
+    async def _ensure_async_session(self):
+        if not self._async_session:
+            self._async_session = mlrun.utils.AsyncClientWithRetry(
+                retry_on_exception=mlrun.mlconf.httpdb.projects.retry_leader_request_on_exception
+                == mlrun.common.schemas.HTTPSessionRetryMode.enabled.value,
+                logger=logger,
+            )
