+++ server/py/framework/utils/clients/iguaziov4.py
@@ -12,8 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-import fastapi
+import typing
 
 import mlrun.common.schemas
 import mlrun.errors
@@ -22,48 +21,33 @@ from framework.utils.clients.base_client import BaseAsyncClient, BaseClient
 
 
 class ClientV4(BaseClient):
-    pass
+    def _generate_auth_info_from_session_verification_response(
+        self,
+        response_headers: typing.Mapping[str, typing.Any],
+        response_body: typing.Mapping[typing.Any, typing.Any],
+    ) -> mlrun.common.schemas.AuthInfo:
+        raise NotImplementedError()
+
+    @property
+    def _verify_session_http_method(self) -> str:
+        return "get"
+
+    def _prepare_request_kwargs(
+        self, session: typing.Optional[str], path: str, *, kwargs: dict
+    ):
+        raise NotImplementedError()
+
+    def _handle_error_response(
+        self,
+        method: str,
+        path: str,
+        response: typing.Any,  # or aiohttp.ClientResponse
+        response_body: dict,
+        error_message: str,
+        kwargs: dict,
+    ) -> None:
+        raise NotImplementedError()
 
 
 class AsyncClient(BaseAsyncClient, ClientV4):
-    async def verify_request_session(
-        self, request: fastapi.Request
-    ) -> mlrun.common.schemas.AuthInfo:
-        """
-        Verifies the session of the incoming request for IG4 environments:
-        Requires either an Authorization header (JWT) or _oauth2_proxy cookie.
-        """
-        authorization = request.headers.get("authorization")
-        oauth2_cookie = request.cookies.get("_oauth2_proxy")
-
-        # Enforce presence of at least one valid auth credential
-        if not authorization and not oauth2_cookie:
-            raise mlrun.errors.MLRunUnauthorizedError(
-                "Missing authentication credentials: expected either Authorization header (JWT) or _oauth2_proxy cookie"
-            )
-
-        user_info = await self._fetch_user_info(request)
-
-        return self._parse_auth_info_from_user_info(
-            user_info, session=authorization or oauth2_cookie
-        )
-
-    async def _fetch_user_info(self, request: fastapi.Request) -> dict:
-        raise NotImplementedError()
-
-    def _parse_auth_info_from_user_info(
-        self, user_info: dict, session: str
-    ) -> mlrun.common.schemas.AuthInfo:
-        username = user_info.get("username")
-        group_ids = user_info.get("groups", [])
-
-        if not username:
-            raise mlrun.errors.MLRunUnauthorizedError(
-                "Received invalid user identity from Iguazio"
-            )
-
-        return mlrun.common.schemas.AuthInfo(
-            username=username,
-            session=session,
-            user_group_ids=group_ids,
-        )
+    pass
